requirements.txt
streamlit
numpy
pillow
pandas
plotly
nltk
scikit-learn